Reorganize directories. Update data/projections.
Reorganized directory structure of repository to consist of: (1) a data
directory, where all data on infectious diseases will be stored; (2) a
model directory, where all of the source code used to implement the
epidemiological models is stored; and (3) a projection directory, where
figures highlighting the model fits to the data are stored.

Updated covid-19 dataset and made new projection.

Cleaned up discussion and comments in code for SI2RD model.